+++ dgisim/src/character/character.py
@@ -48,6 +48,7 @@ __all__ = [
     "MaguuKenki",
     "Mona",
     "Nahida",
+    "Ningguang",
     "Noelle",
     "Qiqi",
     "RhodeiaOfLoch",
@@ -2071,6 +2072,83 @@ class Nahida(Character):
             max_hp=10,
             energy=0,
             max_energy=2,
+            hiddens=stts.Statuses(()),
+            equipments=stts.EquipmentStatuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+
+class Ningguang(Character):
+    _ELEMENT = Element.GEO
+    _WEAPON_TYPE = WeaponType.CATALYST
+    _TALENT_STATUS = stt.StrategicReserveStatus
+    _FACTIONS = frozenset((Faction.LIYUE,))
+
+    _NORMAL_ATTACK_COST = AbstractDices({
+        Element.GEO: 1,
+        Element.ANY: 2,
+    })
+    _ELEMENTAL_SKILL1_COST = AbstractDices({
+        Element.GEO: 3,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDices({
+        Element.GEO: 3,
+    })
+
+    @override
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.GEO,
+            damage=1,
+        )
+
+    @override
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.GEO,
+                damage=2,
+                damage_type=DamageType(elemental_skill=True),
+            ),
+            eft.AddCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.JadeScreenStatus,
+            ),
+        )
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        final_dmg = 6
+        if stt.JadeScreenStatus in game_state.get_player(source.pid).get_combat_statuses():
+            final_dmg += 2
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.GEO,
+                damage=final_dmg,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=3,
             hiddens=stts.Statuses(()),
             equipments=stts.EquipmentStatuses(()),
             statuses=stts.Statuses(()),